1st complete implementation

+++ Money.IO/Money.IO/Record.cs
@@ -5,9 +5,9 @@ namespace Money.IO
     public class Record
     {
         public DateTime DateTime { get; }
-        public float Amount { get; }
+        public double Amount { get; }
 
-        public Record(DateTime dateTime, float amount)
+        public Record(DateTime dateTime, double amount)
         {
             DateTime = dateTime;
             Amount = amount;
